requirements.txt
# Core Dependencies
streamlit==1.29.0
pandas==2.1.3
numpy==1.24.3
plotly==5.18.0
altair==5.2.0

# Database
psycopg2-binary==2.9.9
redshift-connector==2.1.1
sqlalchemy==2.0.23

# AWS
boto3==1.34.5
botocore==1.34.5

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# Data Processing
scikit-learn==1.3.2
scipy==1.11.4
statsmodels==0.14.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
tqdm==4.66.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Logging
loguru==0.7.2

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1